fix: TransferPreset uint8_t→uint32_t 修复虚函数参数垃圾值 + 文件夹传输修复\n\n- 修复 TransferPreset : uint8_t 导致 MSVC x64 虚函数调用参数对齐错误\n  parallelism/speed_limit 传入 createTask 时为垃圾值\n- 添加 Transferring→Completed 状态转换（文件夹传输跳过校验时状态卡住）\n- 修复文件夹传输目标路径：dir_target = dst_path / src_path.filename()\n- 文件夹传输 Step 2 跳过目录哈希，Step 4 跳过目录校验\n- ResumeEngine: pathStemToUtf8 替代 .stem().string() 修复中文路径\n- ResumeEngine: atomicWrite 实现（先写 .tmp 再 rename）\n- ResumeEngine: isSourceFileChanged/invalidateResumeFile 使用 error_code\n- ChunkCompletedCallback: worker 每个 chunk 完成后通知 ResumeEngine\n- recoverFromCrash 增强：加载续传文件任务信息到 active_tasks_\n- GUI: speed_spin_ 限速输入框 + parallelism/speed_limit 传递到 createTask\n- 诊断日志: createTask/executeTaskInner 打印 parallelism/speed_limit 值

+++ Config/ConfigManager.h
@@ -18,7 +18,7 @@ struct SpeedSchedule {
     bool enabled = true;
 };
 
-enum class TransferPreset : uint8_t {
+enum class TransferPreset : uint32_t {
     Fast,
     Secure,
     Balanced
